+++ apps/web/src/hooks/contracts.ts
@@ -1,0 +1,142 @@
+import {
+  createUseReadContract,
+  createUseWriteContract,
+  createUseSimulateContract,
+} from 'wagmi/codegen'
+
+//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+// Counter
+//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+/**
+ * [__View Contract on Sepolia Etherscan__](https://sepolia.etherscan.io/address/0xe789E1d1D0833bC93d76a258aB90DC06674948a5)
+ */
+export const counterAbi = [
+  {
+    type: 'function',
+    inputs: [],
+    name: 'increment',
+    outputs: [],
+    stateMutability: 'nonpayable',
+  },
+  {
+    type: 'function',
+    inputs: [],
+    name: 'number',
+    outputs: [{ name: '', internalType: 'uint256', type: 'uint256' }],
+    stateMutability: 'view',
+  },
+  {
+    type: 'function',
+    inputs: [{ name: 'newNumber', internalType: 'uint256', type: 'uint256' }],
+    name: 'setNumber',
+    outputs: [],
+    stateMutability: 'nonpayable',
+  },
+] as const
+
+/**
+ * [__View Contract on Sepolia Etherscan__](https://sepolia.etherscan.io/address/0xe789E1d1D0833bC93d76a258aB90DC06674948a5)
+ */
+export const counterAddress = {
+  11155111: '0xe789E1d1D0833bC93d76a258aB90DC06674948a5',
+} as const
+
+/**
+ * [__View Contract on Sepolia Etherscan__](https://sepolia.etherscan.io/address/0xe789E1d1D0833bC93d76a258aB90DC06674948a5)
+ */
+export const counterConfig = {
+  address: counterAddress,
+  abi: counterAbi,
+} as const
+
+//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+// React
+//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+/**
+ * Wraps __{@link useReadContract}__ with `abi` set to __{@link counterAbi}__
+ *
+ * [__View Contract on Sepolia Etherscan__](https://sepolia.etherscan.io/address/0xe789E1d1D0833bC93d76a258aB90DC06674948a5)
+ */
+export const useReadCounter = /*#__PURE__*/ createUseReadContract({
+  abi: counterAbi,
+  address: counterAddress,
+})
+
+/**
+ * Wraps __{@link useReadContract}__ with `abi` set to __{@link counterAbi}__ and `functionName` set to `"number"`
+ *
+ * [__View Contract on Sepolia Etherscan__](https://sepolia.etherscan.io/address/0xe789E1d1D0833bC93d76a258aB90DC06674948a5)
+ */
+export const useReadCounterNumber = /*#__PURE__*/ createUseReadContract({
+  abi: counterAbi,
+  address: counterAddress,
+  functionName: 'number',
+})
+
+/**
+ * Wraps __{@link useWriteContract}__ with `abi` set to __{@link counterAbi}__
+ *
+ * [__View Contract on Sepolia Etherscan__](https://sepolia.etherscan.io/address/0xe789E1d1D0833bC93d76a258aB90DC06674948a5)
+ */
+export const useWriteCounter = /*#__PURE__*/ createUseWriteContract({
+  abi: counterAbi,
+  address: counterAddress,
+})
+
+/**
+ * Wraps __{@link useWriteContract}__ with `abi` set to __{@link counterAbi}__ and `functionName` set to `"increment"`
+ *
+ * [__View Contract on Sepolia Etherscan__](https://sepolia.etherscan.io/address/0xe789E1d1D0833bC93d76a258aB90DC06674948a5)
+ */
+export const useWriteCounterIncrement = /*#__PURE__*/ createUseWriteContract({
+  abi: counterAbi,
+  address: counterAddress,
+  functionName: 'increment',
+})
+
+/**
+ * Wraps __{@link useWriteContract}__ with `abi` set to __{@link counterAbi}__ and `functionName` set to `"setNumber"`
+ *
+ * [__View Contract on Sepolia Etherscan__](https://sepolia.etherscan.io/address/0xe789E1d1D0833bC93d76a258aB90DC06674948a5)
+ */
+export const useWriteCounterSetNumber = /*#__PURE__*/ createUseWriteContract({
+  abi: counterAbi,
+  address: counterAddress,
+  functionName: 'setNumber',
+})
+
+/**
+ * Wraps __{@link useSimulateContract}__ with `abi` set to __{@link counterAbi}__
+ *
+ * [__View Contract on Sepolia Etherscan__](https://sepolia.etherscan.io/address/0xe789E1d1D0833bC93d76a258aB90DC06674948a5)
+ */
+export const useSimulateCounter = /*#__PURE__*/ createUseSimulateContract({
+  abi: counterAbi,
+  address: counterAddress,
+})
+
+/**
+ * Wraps __{@link useSimulateContract}__ with `abi` set to __{@link counterAbi}__ and `functionName` set to `"increment"`
+ *
+ * [__View Contract on Sepolia Etherscan__](https://sepolia.etherscan.io/address/0xe789E1d1D0833bC93d76a258aB90DC06674948a5)
+ */
+export const useSimulateCounterIncrement =
+  /*#__PURE__*/ createUseSimulateContract({
+    abi: counterAbi,
+    address: counterAddress,
+    functionName: 'increment',
+  })
+
+/**
+ * Wraps __{@link useSimulateContract}__ with `abi` set to __{@link counterAbi}__ and `functionName` set to `"setNumber"`
+ *
+ * [__View Contract on Sepolia Etherscan__](https://sepolia.etherscan.io/address/0xe789E1d1D0833bC93d76a258aB90DC06674948a5)
+ */
+export const useSimulateCounterSetNumber =
+  /*#__PURE__*/ createUseSimulateContract({
+    abi: counterAbi,
+    address: counterAddress,
+    functionName: 'setNumber',
+  })