+++ apps/web/src/main.tsx
@@ -1,18 +1,18 @@
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { WagmiProvider } from "wagmi";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 
 import { ThemeProvider } from "@/components/theme-provider";
 
-import App from "./App.tsx";
+import App from "./app.tsx";
 import { config } from "./wagmi.ts";
 
 import "@repo/ui/index.css";
 
 const queryClient = new QueryClient();
 
-createRoot(document.getElementById("root")!).render(
+createRoot(document.querySelector("#root")!).render(
   <StrictMode>
     <ThemeProvider>
       <WagmiProvider config={config}>